# Calculate filesystem node's hash lazily.