@pjbgf 2023-08-06
@candid82 thanks for your PR. Do you mind providing a [benchstat](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat) of the performance improvements please?

Also, re-word the commit message to align with the [contribution guidelines](https://github.com/go-git/go-git/blob/master/CONTRIBUTING.md#format-of-the-commit-message).

> A better fix (instead of or in addition to this one) would be to avoid traversing ignored files in the first place. However, such change seem to be more intrusive and will require much deeper understanding of the codebase.

Happy for this to be a follow-up PR.