@candid82 2023-08-22
@pjbgf Below please find benchstat results for one of my local repositories with sizable node_modules directory. Obviously, the results depend on the size of git repo and the number of ignored files. 
```
> benchstat ~/old.txt  ~/new.txt
goos: darwin
goarch: amd64
pkg: github.com/go-git/go-git/v5
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
          │ /Users/candid/old.txt │       /Users/candid/new.txt        │
          │        sec/op         │   sec/op    vs base                │
Status-16              3.875 ± 1%   1.392 ± 3%  -64.07% (p=0.000 n=10)
```

The benchmark code:
```
func BenchmarkStatus(b *testing.B) {
	repo, _ := PlainOpen("<path to repo>")
	worktree, _ := repo.Worktree()
	for i := 0; i < b.N; i++ {
		s, _ := worktree.Status()
		if len(s) > 0 {
			b.Errorf("error!")
		}
	}
}
```

I also adjusted the commit message.