@pjbgf 2023-09-30
Thank you for providing the benchstat and rewording the commit. 

Doing another review I noticed that this change could introduce an issue by which once the `Hash()` is called and the hash is calculated, changes to the underlying `fs` will not trigger a new hash recalculation. By looking for the places where this is used, I see no red flags that this would be a problem. Please add some comments in the `Hash()` func to that effect.

Otherwise, LGTM.