@codablock 2024-01-08
@candid82 @pjbgf Any updates on this PR? If there is lack of time on @candid82's side, I'm also happy to provide the requested comments/docs after this PR is merged.

Merging this PR would alleviate a lot of pain in regard to https://github.com/go-git/go-git/issues/181. On my side, it would help me avoid a temporary workaround that would exec-out into `git status`, so I'm very happy to instead help here :)