@codablock 2024-01-10
@pjbgf As requested in Slack, I have tested this PR locally (by replacing `github.com/go-git/go-git/v5` with my local version inside go.mod) and can confirm that `Status()` performance gets improved a lot. I used the notorious `node_modules` to test things out and Status() runtime has been reduced to below a second from previously 5-10 seconds.

If you merge this PR, I will follow up with a PR that adds the docs/comments.