@codablock 2024-01-10
Here is the PR with the requested comment update: https://github.com/go-git/go-git/pull/992